@@ -13,8 +13,8 @@ group 'io.github.mfvanek'
 version '0.6.3-SNAPSHOT'
 
 java {
-    sourceCompatibility = JavaVersion.VERSION_1_8
-    targetCompatibility = JavaVersion.VERSION_1_8
+    sourceCompatibility = JavaVersion.VERSION_11
+    targetCompatibility = JavaVersion.VERSION_11
 }
 
 repositories {
@@ -23,7 +23,7 @@ repositories {
 }
 
 dependencies {
-    implementation 'ch.qos.logback:logback-classic:1.3.1'
+    implementation 'ch.qos.logback:logback-classic:1.4.1'
     implementation 'io.github.mfvanek:pg-index-health:0.6.2'
     implementation 'org.postgresql:postgresql:42.5.0'
     implementation 'com.google.code.findbugs:jsr305:3.0.2'
@@ -42,6 +42,7 @@ dependencies {
     testImplementation 'org.mockito:mockito-core:4.8.0'
 
     pitest 'it.mulders.stryker:pit-dashboard-reporter:0.1.2'
+    checkstyle 'com.thomasjensen.checkstyle.addons:checkstyle-addons:7.0.1'
 }
 
 test {
